Fix colorscheme workaround
For some reason on some versions of GVim, the colorscheme initially
loaded in the GUI is messed up until you reload it once. Fix the code to
work around this issue: the original code will change back the
colorscheme to the one in the .vimrc file if you change your colorscheme
after you start the program, which is unnecessary since the autocmd used
to work around the issue can just be undefined after it's used once.
Also allow specification of the background type (i.e., "light" or
"dark").