Attempt to unify commands
I believe that, within the place in M syntax where a command needs to be
placed (i.e., right after whitespace at the start of a line or after
another complete statement plus whitespace), it should be possible to
multiplex commands with different types of arguments (i.e. do, goto,
new, write, etc.) into one mCmd group. I've attempted to do this here,
but it matches on generic commands and fails on the special cases.